+++ test/pleroma/gun/connection_pool_test.exs
@@ -19,7 +19,6 @@ defmodule Pleroma.Gun.ConnectionPoolTest do
     :ok
   end
 
-  setup :set_mox_from_context
   setup :gun_mock
 
   test "gives the same connection to 2 concurrent requests" do